--- src/components/app/styles/AppMode.module.css
@@ -1,38 +1,0 @@
-.mapPosition {
-    position: absolute;
-}
-
-.mapPosition > div {
-    height: 100%;
-}
-
-.mapDownload {
-    border: var(--spacers-dp24) solid #fff
-}
-
-.mapDownload :global(.dhis2-map-timeline) {
-    display: none;
-}
-
-.mapDownload :global(.dhis2-map-period) {
-    bottom: 10px !important;
-} 
-
-.mapDownload :global(.maplibregl-ctrl-group:not(:empty)) {
-    box-shadow: none;
-}
-
-.mapDownload > div > div:first-child {
-    border: 1px solid #ccc;
-}
-
-.downloadPanelOpen  > div > div:first-child {
-    width: calc(100% - 380px);
-}
-
-.mapContainer {
-    height: 100%;
-    width: 100%;
-    position: relative;
-    overflow: hidden;
-}
