+++ src/components/download/styles/NorthArrowPosition.module.css
